LayoutTests/crypto/subtle/ec-import-raw-key-export-spki-key-expected.txt

Test importing an ECDH raw key and then export it in SPKI format

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
PASS bytesToHexString(publicKey) is expectedSpkiKey
PASS successfullyParsed is true

TEST COMPLETE
